requirements.txt
cachetools==2.0.0
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.2
gcloud==0.18.3
google-auth==1.0.2
google-cloud-bigquery==0.26.0
google-cloud-core==0.26.0
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
httplib2==0.10.3
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
monotonic==1.3
numpy==1.12.1
oauth2client==4.1.2
oic==0.10.0.0
pandas==0.20.2
plotly==2.0.6
python-dateutil==2.6.1
pytz==2017.2
protobuf==3.3.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
PyJWT==1.4.2
requests==2.18.3
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.1.14
tenacity==4.4.0
urllib3==1.22
Werkzeug==0.12.2